Fix for https://bugs.eclipse.org/bugs/show_bug.cgi?id=149224.
When adding a ['foo'] style property in a content assist,the proposal strategy now checks to see if the preceding char is a '.' and if so, it will replace it as part of the proposal if that proposal is selected by the user.

Regression test coverage also added.

Also made a small thread sync change in JSPModelProcessor to fix a possible race condition in refresh().