Bug 569660 Lazy compute additional proposal info
Instead of computing the additional info of each content assist proposal
before displaying the content assist popup, do it lazy.

Remove parameter additionalText from the method createProposal of the
interface IContentProposalProvider. Therefore the additionalText must
not be provided when creating the proposal.

The new class LazyAdditionalTextCompletionProposal uses the interface
IDescibedInDetail to the get the detailed description and provide is as
additional info. This code was duplicated before in a lot of classes
that provided proposals.

Change-Id: I18bcf3eb86bda3f129538f31f7abf9dfb6e0f310
Signed-off-by: Reto Weiss <reto.weiss@axonivy.com>
Also-by: Andreas Rusch <andreas.rusch@axonivy.com>
Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=569660